refactor: move `extract_environment_from_headers` to utils.py
